+++ CI/installation/install_dep_umoci.bash
@@ -4,8 +4,8 @@ pwd_bak=$PWD
 
 # Install umoci
 cd /tmp && \
-    wget -O umoci.amd64 https://github.com/opencontainers/umoci/releases/download/v0.4.7/umoci.amd64 && \
-    chmod 755 umoci.amd64 && \
-    sudo mv umoci.amd64 /usr/local/bin/ && \
-    sudo chown root:root /usr/local/bin/umoci.amd64 && \
+    wget -O umoci https://github.com/opencontainers/umoci/releases/download/v0.4.7/umoci.amd64 && \
+    chmod 755 umoci && \
+    sudo mv umoci /usr/local/bin/ && \
+    sudo chown root:root /usr/local/bin/umoci && \
     cd ${pwd_bak}