+++ CI/installation/install_dependencies.bash
@@ -8,7 +8,7 @@ ${base_dir}/install_dep_boost.bash
 ${base_dir}/install_dep_rapidjson.bash
 ${base_dir}/install_dep_runc.bash
 ${base_dir}/install_dep_tini.bash
-${base_dir}/install_dep_skopeo.bash
+${base_dir}/install_dep_skopeo_static.bash
 ${base_dir}/install_dep_umoci.bash
 
 # Note Dropbear and cpputest are also dependencies, for now managed by cmake.