+++ CI/installation/install_packages_debian:10.sh
@@ -1,1 +1,0 @@
-install_packages_ubuntu:20.04.sh